Replace time periods select by buttons group
